+++ test/test_mhrwtasks.cxx
@@ -63,7 +63,7 @@ BOOST_AUTO_TEST_SUITE(cdatabase)
 
 BOOST_AUTO_TEST_CASE(constr)
 {
-  Tomographer::MHRWTasks::CDataBase<> cdata(Tomographer::MHRWParams<int, double>(0.1, 128, 50, 500), 100);
+  Tomographer::MHRWTasks::CDataBase<> cdata(Tomographer::MHRWParams<double, int>(0.1, 128, 50, 500), 100);
   BOOST_CHECK_EQUAL(cdata.base_seed, 100);
   MY_BOOST_CHECK_FLOATS_EQUAL(cdata.mhrw_params.mhwalker_params, 0.1, tol);
   BOOST_CHECK_EQUAL(cdata.mhrw_params.n_sweep, 128);
@@ -73,7 +73,7 @@ BOOST_AUTO_TEST_CASE(constr)
 
 BOOST_AUTO_TEST_CASE(get_task_inputs)
 {
-  const Tomographer::MHRWTasks::CDataBase<> cdata(Tomographer::MHRWParams<int, double>(0.1, 128, 50, 500), 100);
+  const Tomographer::MHRWTasks::CDataBase<> cdata(Tomographer::MHRWParams<double, int>(0.1, 128, 50, 500), 100);
 
   BOOST_MESSAGE( cdata.getBasicCDataMHRWInfo() );
 
@@ -100,7 +100,7 @@ struct MyStatsResultType {
   MyStatsResultType() : a(0), b(1) { }
   int a, b;
 };
-typedef Tomographer::MHRWParams<int,double> MHRWParamsType;
+typedef Tomographer::MHRWParams<double,int> MHRWParamsType;
 
 BOOST_AUTO_TEST_CASE(instantiate_1)
 {
